refactor: restructure authentication module app
