+++ backend/app/users/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from app.authentication.views.auth_views import UserDetailView
+from app.auth.views.auth_views import UserDetailView
 
 from app.users.views import (
     UserListView,
